+++ assessment-api/assessment-service/app/modules/AssessmentModule.scala
@@ -15,7 +15,6 @@ class AssessmentModule extends AbstractModule with AkkaGuiceSupport {
         bindActor(classOf[QuestionSetActor], ActorNames.QUESTION_SET_ACTOR)
         bindActor(classOf[org.sunbird.v5.actors.QuestionActor], ActorNames.QUESTION_V5_ACTOR)
         bindActor(classOf[org.sunbird.v5.actors.QuestionSetActor], ActorNames.QUESTION_SET_V5_ACTOR)
-        bindActor(classOf[org.sunbird.v5.actors.QuestionSetAssessActor], ActorNames.QUESTION_SET_ASSESS_ACTOR)
         println("Initialized application actors for assessment-service")
     }
 }
