+++ assessment-api/assessment-service/app/utils/ActorNames.scala
@@ -8,6 +8,5 @@ object ActorNames {
     final val QUESTION_SET_ACTOR = "questionSetActor"
     final val QUESTION_V5_ACTOR = "questionV5Actor"
     final val QUESTION_SET_V5_ACTOR = "questionSetV5Actor"
-    final val QUESTION_SET_ASSESS_ACTOR="questionSetAssessActor"
 
 }
